+++ core/governance/engine.go
@@ -24,6 +24,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/pkg/errors"
+
 	"code.vegaprotocol.io/vega/core/assets"
 	"code.vegaprotocol.io/vega/core/events"
 	"code.vegaprotocol.io/vega/core/netparams"
@@ -32,7 +34,6 @@ import (
 	vgcrypto "code.vegaprotocol.io/vega/libs/crypto"
 	"code.vegaprotocol.io/vega/libs/num"
 	"code.vegaprotocol.io/vega/logging"
-	"github.com/pkg/errors"
 )
 
 var (
@@ -286,7 +287,7 @@ func (e *Engine) preEnactProposal(ctx context.Context, p *proposal) (te *ToEnact
 	case types.ProposalTermsTypeUpdateMarketState:
 		te.msu = &ToEnactMarketStateUpdate{}
 	case types.ProposalTermsTypeUpdateReferralProgram:
-		te.referralProgram = updatedReferralProgramFromProposal(p)
+		te.referralProgramChanges = updatedReferralProgramFromProposal(p)
 	case types.ProposalTermsTypeUpdateVolumeDiscountProgram:
 		te.volumeDiscountProgram = updatedVolumeDiscountProgramFromProposal(p)
 	}