+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [9258](https://github.com/vegaprotocol/vega/issues/9258) - Rework `GetReferralSetStats` endpoint.
 - [9258](https://github.com/vegaprotocol/vega/issues/9258) - Change HTTP endpoint from `/volume-discount-stats` to `/volume-discount-program/stats`.
 - [9258](https://github.com/vegaprotocol/vega/issues/9258) - Change HTTP endpoint from `/referral-sets/stats/{id}` to `/referral-sets/{id}/stats`.
+- [9719](https://github.com/vegaprotocol/vega/issues/9719) - Remove unnecessary fields from referral and volume discount program proposals.
 
 ### 🗑️ Deprecation